requirements.txt
streamlit>=1.28.0
pymysql>=1.0.0
PyYAML>=5.4.0
pandas>=1.3.0
requests>=2.25.0
dbutils>=3.0.0
schedule>=1.1.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0
ipywidgets>=7.6.0